feat: 🎸 Refactor types into module specific files
